% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_gosa_data}
\alias{download_gosa_data}
\title{Download GOSA enrollment data}
\usage{
download_gosa_data(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with enrollment data frame
}
\description{
Downloads enrollment data from GOSA's downloadable data repository.
Uses Enrollment by Subgroup Metrics for demographics and special populations.
}
\keyword{internal}
