% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_gosa_subgroup}
\alias{download_gosa_subgroup}
\title{Download GOSA Enrollment by Subgroup data}
\usage{
download_gosa_subgroup(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with subgroup enrollment data
}
\description{
Downloads the Enrollment_by_Subgroup_Metrics CSV from GOSA.
This file contains demographic breakdowns and special population counts.
}
\keyword{internal}
