% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_format_era}
\alias{get_format_era}
\title{Get format era for a given year}
\usage{
get_format_era(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character string indicating era (always "gosa" for supported years)
}
\description{
Returns the data format era for processing. All data comes from GOSA.
}
\keyword{internal}
