% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data for Georgia}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024). Valid range: 2011-2024}
}
\value{
List with enrollment data frame(s)
}
\description{
Downloads enrollment data from the Governor's Office of Student Achievement
(GOSA) downloadable data repository.
}
\keyword{internal}
