% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{merge_gosa_data}
\alias{merge_gosa_data}
\title{Merge GOSA subgroup and grade data}
\usage{
merge_gosa_data(subgroup_df, grade_df)
}
\arguments{
\item{subgroup_df}{Subgroup enrollment data}

\item{grade_df}{Grade-level enrollment data}
}
\value{
Merged data frame
}
\description{
Combines the subgroup demographics with grade-level enrollment.
}
\keyword{internal}
